AG-14318 Add a snapshot-based lint-change-detection.test.ts implementation
