+++ packages/ag-charts-community/tsconfig.lib.json
@@ -5,5 +5,5 @@
     "types": ["node"]
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts", "src/**/test/**"]
+  "exclude": ["**/*.data.ts", "jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts", "src/**/test/**"]
 }
