+++ packages/ag-charts-community/tsconfig.types.test.json
@@ -5,5 +5,6 @@
     "emitDeclarationOnly": true,
     "types": ["node"]
   },
-  "include": ["jest.config.ts", "jest.setup.ts", "src/main-test.ts", "**/*.test.ts", "**/test/**"]
+  "include": ["**/*.data.ts", "jest.config.ts", "jest.setup.ts", "src/main-test.ts", "**/*.test.ts", "**/test/**"],
+  "exclude": ["**/*.data.ts"]
 }
