+++ packages/ag-charts-community/tsconfig.watch.json
@@ -4,5 +4,6 @@
     "strict": true,
     "noUnusedParameters": false,
     "noUnusedLocals": false
-  }
+  },
+  "exclude": ["**/*.data.ts"]
 }
